Fix RSocket Package Tangle
Issue gh-7360
