+++ config/src/main/java/org/springframework/security/config/annotation/rsocket/RSocketSecurity.java
@@ -28,18 +28,18 @@ import org.springframework.security.authorization.ReactiveAuthorizationManager;
 import org.springframework.security.config.Customizer;
 import org.springframework.security.oauth2.jwt.ReactiveJwtDecoder;
 import org.springframework.security.oauth2.server.resource.authentication.JwtReactiveAuthenticationManager;
-import org.springframework.security.rsocket.interceptor.PayloadInterceptor;
-import org.springframework.security.rsocket.interceptor.PayloadSocketAcceptorInterceptor;
-import org.springframework.security.rsocket.interceptor.authentication.AnonymousPayloadInterceptor;
-import org.springframework.security.rsocket.interceptor.authentication.AuthenticationPayloadInterceptor;
-import org.springframework.security.rsocket.interceptor.authentication.BearerPayloadExchangeConverter;
-import org.springframework.security.rsocket.interceptor.authorization.AuthorizationPayloadInterceptor;
-import org.springframework.security.rsocket.interceptor.authorization.PayloadExchangeMatcherReactiveAuthorizationManager;
-import org.springframework.security.rsocket.util.PayloadExchangeAuthorizationContext;
-import org.springframework.security.rsocket.util.PayloadExchangeMatcher;
-import org.springframework.security.rsocket.util.PayloadExchangeMatcherEntry;
-import org.springframework.security.rsocket.util.PayloadExchangeMatchers;
-import org.springframework.security.rsocket.util.RoutePayloadExchangeMatcher;
+import org.springframework.security.rsocket.PayloadInterceptor;
+import org.springframework.security.rsocket.PayloadSocketAcceptorInterceptor;
+import org.springframework.security.rsocket.authentication.AnonymousPayloadInterceptor;
+import org.springframework.security.rsocket.authentication.AuthenticationPayloadInterceptor;
+import org.springframework.security.rsocket.authentication.BearerPayloadExchangeConverter;
+import org.springframework.security.rsocket.authorization.AuthorizationPayloadInterceptor;
+import org.springframework.security.rsocket.authorization.PayloadExchangeMatcherReactiveAuthorizationManager;
+import org.springframework.security.rsocket.util.matcher.PayloadExchangeAuthorizationContext;
+import org.springframework.security.rsocket.util.matcher.PayloadExchangeMatcher;
+import org.springframework.security.rsocket.util.matcher.PayloadExchangeMatcherEntry;
+import org.springframework.security.rsocket.util.matcher.PayloadExchangeMatchers;
+import org.springframework.security.rsocket.util.matcher.RoutePayloadExchangeMatcher;
 import reactor.core.publisher.Mono;
 
 import java.util.ArrayList;
