+++ rsocket/src/main/java/org/springframework/security/rsocket/ContextPayloadInterceptorChain.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.security.rsocket.interceptor;
+package org.springframework.security.rsocket;
 
 import reactor.core.publisher.Mono;
 import reactor.util.context.Context;
