+++ rsocket/src/main/java/org/springframework/security/rsocket/PayloadExchange.java
@@ -14,21 +14,23 @@
  * limitations under the License.
  */
 
-package org.springframework.security.rsocket.interceptor;
+package org.springframework.security.rsocket;
 
-import reactor.core.publisher.Mono;
+import io.rsocket.Payload;
+import org.springframework.util.MimeType;
 
 /**
- * Contract to allow a {@link PayloadInterceptor} to delegate to the next in the chain.
- *  *
+ * Contract for a Payload interaction.
+ *
  * @author Rob Winch
  * @since 5.2
  */
-public interface PayloadInterceptorChain {
-	/**
-	 * Process the payload exchange.
-	 * @param exchange the current server exchange
-	 * @return {@code Mono<Void>} to indicate when request processing is complete
-	 */
-	Mono<Void> next(PayloadExchange exchange);
+public interface PayloadExchange {
+	PayloadExchangeType getType();
+
+	Payload getPayload();
+
+	MimeType getDataMimeType();
+
+	MimeType getMetadataMimeType();
 }
