+++ rsocket/src/main/java/org/springframework/security/rsocket/PayloadInterceptor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.security.rsocket.interceptor;
+package org.springframework.security.rsocket;
 
 import reactor.core.publisher.Mono;
