+++ rsocket/src/main/java/org/springframework/security/rsocket/PayloadSocketAcceptor.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.springframework.security.rsocket.interceptor;
+package org.springframework.security.rsocket;
 
 import io.rsocket.ConnectionSetupPayload;
 import io.rsocket.Payload;
