+++ rsocket/src/main/java/org/springframework/security/rsocket/authentication/BearerPayloadExchangeConverter.java
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package org.springframework.security.rsocket.interceptor.authentication;
+package org.springframework.security.rsocket.authentication;
 
 import io.netty.buffer.ByteBuf;
 import io.rsocket.metadata.CompositeMetadata;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.oauth2.server.resource.BearerTokenAuthenticationToken;
-import org.springframework.security.rsocket.interceptor.PayloadExchange;
+import org.springframework.security.rsocket.PayloadExchange;
 import org.springframework.security.rsocket.metadata.BearerTokenMetadata;
 import reactor.core.publisher.Mono;
