+++ rsocket/src/main/java/org/springframework/security/rsocket/util/matcher/RoutePayloadExchangeMatcher.java
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package org.springframework.security.rsocket.util;
+package org.springframework.security.rsocket.util.matcher;
 
 import org.springframework.messaging.rsocket.MetadataExtractor;
-import org.springframework.security.rsocket.interceptor.PayloadExchange;
+import org.springframework.security.rsocket.PayloadExchange;
 import org.springframework.util.Assert;
 import org.springframework.util.RouteMatcher;
 import reactor.core.publisher.Mono;
